Add files via upload
Cancer is a group of closely linked diseases that all
involve unchecked cellular growth and reproduction,
rather than being one unique illness. Cancer kills almost 8
million people a year, making it the top cause of death in
the industrialised world and the second biggest cause in
the developing world. Machine learning is a branch of
artificial intelligence that makes use of a range of
statistical, probabilistic, and optimization methods to
enable computers to "learn" from previous experiences
and to recognise difficult-to-find patterns in vast, noisy, or
complex data sets.Medical applications, particularly those
that rely on intricate proteomic and genomic data, are
especially well suited for this capability. As a result,
machine learning is frequently employed in the detection
and diagnosis of cancer.An improved comprehension of
risk variables can aid in the prevention of lung cancer.
The key to increasing the likelihood of survival is early
detection using machine learning techniques, and if we
can use this to make the diagnostic process for
radiologists more effective and efficient, it will be a
significant step towards improving early detection.More
than 200 different types of cancer can be identified using
molecular classification techniques, and in order for the
patient to receive the right treatment, the practitioner must
determine the cancer type as precisely as possible. Thus,
molecular diagnostic techniques are required in order to
categorise tumour subtypes accurately. Thousands of
genes are represented in the dozens of samples that make
up gene expression data. Thus, it is difficult to classify
such samples in a precise and efficient manner. In-depth
and accurate categorization models have been created
using a wide variety of machine learning approaches.
Using Random Forest, K-nearest Neighbour, Support
Vector Machines, Logistic Regression, and Decision
Trees, this study suggests an ensemble learning-based
strategy.